backend/requirements.txt
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.0.1
PyMuPDF==1.23.5
pdfplumber==0.9.0
